bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["SosisterRapStar <timoshenkovvanya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.13.0"
pydantic-settings = "^2.5.2"
aiohttp = "^3.10.5"
ruff = "^0.6.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
